RUBY-1033: Explain command passing
