--- lib/mongo/operation/commands/explain/result.rb
@@ -1,31 +1,0 @@
-# Copyright (C) 2015 MongoDB, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-module Mongo
-  module Operation
-    class Explain
-
-      # Defines custom behaviour of results in find command.
-      #
-      # @since 2.2.0
-      class Result < Operation::Result
-
-        def documents
-          p reply.documents
-          super
-        end
-      end
-    end
-  end
-end
